♻️ Migrate the prompt provider which has hot loading feature into another module *provider* to converge the relative logic.
